requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
fastapi==0.129.0
google-auth==2.49.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
passlib==1.7.4
pip==24.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
SQLAlchemy==2.0.46
starlette==0.52.1
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
websockets==16.0
